refactor: point layout
