--- packages/mesh/app/point/subscribe-stories/layout.tsx
@@ -1,50 +1,0 @@
-'use client'
-import Link from 'next/link'
-import { useRouter } from 'next/navigation'
-
-import Icon from '@/components/icon'
-import LayoutTemplate from '@/components/layout-template'
-
-export default function SubscribeStoriesLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  const router = useRouter()
-
-  const backToPreviousPage = () => {
-    router.back()
-  }
-
-  return (
-    <LayoutTemplate
-      type="default"
-      customStyle={{
-        background: 'bg-white sm:bg-multi-layer-light',
-        restrictMainWidth: true,
-        footer: 'hidden sm:block',
-      }}
-      navigation={{
-        leftButtons: [
-          {
-            type: 'icon',
-            icon: 'icon-chevron-left',
-            onClick: backToPreviousPage,
-          },
-        ],
-        title: '訂閱中文章',
-        rightButtons: [],
-      }}
-    >
-      {/* TODO: use shared pc navigation component */}
-      <div className="hidden h-15 w-full flex-row items-center border-b bg-white sm:flex">
-        <Link href={'/point'}>
-          <Icon iconName="icon-chevron-left" size="m" className="ml-5" />
-        </Link>
-        <h2 className="list-title mx-auto sm:ml-6">訂閱中文章</h2>
-        <div className="size-5 px-5"></div>
-      </div>
-      {children}
-    </LayoutTemplate>
-  )
-}
